Rebased on adapter test for sqlite
